Medical/requirements.txt
streamlit
agno
groq
python-dotenv
pytesseract
Pillow
